requirements.txt
Flask==1.0
Flask-RESTful==0.3.6
nose_parameterized==0.6.0
nose==1.3.7
coveralls==1.2.0
